+++ app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".SettingsActivity"
+            android:name=".settings.SettingsActivity"
             android:screenOrientation="landscape" />
     </application>
