--- app/src/main/java/com/example/eyespeak/KeyboardView.kt
@@ -1,18 +1,0 @@
-package com.example.eyespeak
-
-import android.content.Context
-import android.graphics.Canvas
-import android.graphics.Color
-import android.graphics.Paint
-import android.graphics.Rect
-import android.util.AttributeSet
-import android.util.Log
-import android.view.View
-
-class KeyboardView(context: Context, attrs: AttributeSet): View(context, attrs) {
-
-    override fun onDraw(canvas: Canvas?) {
-        super.onDraw(canvas)
-    }
-
-}
