+++ app/src/main/java/com/example/eyespeak/MainActivity.kt
@@ -4,6 +4,10 @@ import android.os.Bundle
 import android.view.View
 import android.view.ViewGroup
 import androidx.core.view.children
+import com.example.eyespeak.customviews.PauseButton
+import com.example.eyespeak.customviews.PointerListener
+import com.example.eyespeak.customviews.PointerView
+import com.example.eyespeak.utils.VocableTextToSpeech
 import kotlinx.android.synthetic.main.activity_main.*
