+++ app/src/main/java/com/example/eyespeak/customviews/CategoryButton.kt
@@ -1,9 +1,10 @@
-package com.example.eyespeak
+package com.example.eyespeak.customviews
 
 import android.content.Context
 import android.speech.tts.TextToSpeech
 import android.util.AttributeSet
 import androidx.appcompat.widget.AppCompatRadioButton
+import com.example.eyespeak.utils.VocableTextToSpeech
 import kotlinx.coroutines.*
 
 /**
@@ -13,7 +14,8 @@ class CategoryButton @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
     defStyle: Int = 0
-) : AppCompatRadioButton(context, attrs, defStyle), PointerListener {
+) : AppCompatRadioButton(context, attrs, defStyle),
+    PointerListener {
 
     companion object {
         private const val DEFAULT_TTS_TIMEOUT = 2000L
