+++ app/src/main/java/com/example/eyespeak/customviews/PointerListener.kt
@@ -1,4 +1,4 @@
-package com.example.eyespeak
+package com.example.eyespeak.customviews
 
 interface PointerListener {
