+++ app/src/main/java/com/example/eyespeak/facetracking/FaceTrackFragment.kt
@@ -1,4 +1,4 @@
-package com.example.eyespeak
+package com.example.eyespeak.facetracking
 
 import android.os.Bundle
 import androidx.lifecycle.Observer
