+++ app/src/main/res/layout/activity_settings.xml
@@ -8,7 +8,7 @@
 
     <fragment
         android:id="@+id/face_fragment"
-        android:name="com.example.eyespeak.FaceTrackFragment"
+        android:name="com.example.eyespeak.facetracking.FaceTrackFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
@@ -17,7 +17,7 @@
         android:layout_height="match_parent"
         android:background="@color/colorPrimaryDark" />
 
-    <com.example.eyespeak.PauseButton
+    <com.example.eyespeak.customviews.PauseButton
         android:id="@+id/settings_pause_button"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -31,7 +31,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <com.example.eyespeak.CloseButton
+    <com.example.eyespeak.customviews.CloseButton
         android:id="@+id/settings_close_button"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -60,7 +60,7 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_chainStyle="packed" />
 
-    <com.example.eyespeak.ActionButton
+    <com.example.eyespeak.customviews.ActionButton
         android:id="@+id/privacy_policy_button"
         android:layout_width="0dp"
         android:layout_height="144dp"
@@ -79,7 +79,7 @@
         app:layout_constraintTop_toBottomOf="@+id/settings_title"
         app:layout_constraintVertical_chainStyle="packed" />
 
-    <com.example.eyespeak.ActionButton
+    <com.example.eyespeak.customviews.ActionButton
         android:id="@+id/contact_devs_button"
         android:layout_width="0dp"
         android:layout_height="144dp"
@@ -111,7 +111,7 @@
         app:layout_constraintTop_toBottomOf="@+id/contact_devs_button"
         app:layout_constraintVertical_chainStyle="packed" />
 
-    <com.example.eyespeak.PointerView
+    <com.example.eyespeak.customviews.PointerView
         android:id="@+id/pointer_view"
         android:layout_width="32dp"
         android:layout_height="32dp"
